Add files via upload
Please first look at the Solving_Better_Continous_Bin(), it solves the problem in 50 seconds.
I think Solving_Best_Continous_Bin() should be better and faster due to less IntVar, but it doesnt. It is slower.
Solving_Discrete_Bins() can solve the problem in 10 seconds and its approach is exactly like your MIP in the EJOR article.
I deactivated each method by /* in the first and */ in the last line of each method. For activate each method, please remove them.